[programming_examples/llms] Make prefill scratch args registry-driven (MHA-safe); drop SmolLM2 fork
run_transformer_block / preload_prefill_weights in llama32_1b_prefill hardcoded
a GQA scratch-arg assumption (Q fused-cast -> 1 f32 C-scratch at arg13; K/V drain).
At kv_dim == emb_dim (pure MHA, e.g. SmolLM2-1.7B) the registry resolves K/V to
fused-cast too, so the compiled ELF declares 3 scratch args (13,14,15) but the
caller passed only 1 -- K/V read unallocated scratch and produced zero output.

Make the scratch-arg construction registry-driven via a new _rms_scratch_specs
helper that mirrors build_rms_gemms_rope_module's own per-shape
gemm_registry_config lookup (same Q->K->V order). GQA stays bit-identical
(1 scratch, arg13); MHA correctly gets 3. The _arg_cache rms entry becomes a
(args, scratch_inter) tuple so the intermediate-index set tracks the scratch count.

This was previously worked around by a SmolLM2-only fork
(smollm2_1_7b_prefill.py) that monkeypatched both functions. With the shared
functions now MHA-safe, delete the fork and its monkeypatch -- SmolLM2's
inference + verify_adapter import the shared llama32_1b_prefill directly.

Verified on NPU2 (make verify, 2 prompts x 32 tokens, k=5): llama32_1b PASS,
llama32_1b_int4 PASS (both GQA, regression check), smollm2_1_7b PASS (MHA).
