+++ programming_examples/llms/llama32_1b/llama32_1b_prefill.py
@@ -76,6 +76,33 @@ def _rms_gemms_rope_run_backend():
     return {**_base, "runtime_loop_tiling_sizes": [2, 2]}
 
 
+def _rms_scratch_specs(seq_len, emb_dim, kv_dim):
+    """Registry-driven f32 C-scratch args for the Q/K/V GEMMs of rms_gemms_rope,
+    in builder order (Q, K, V). Returns (list_of_scratch_arrays, set_of_indices).
+
+    One scratch array per GEMM whose registry method is fused-cast; indices start
+    at 13. This MUST mirror build_rms_gemms_rope_module's own per-shape
+    gemm_registry_config lookup (same shapes, same Q->K->V order) so the args we
+    pass match the scratch args the compiled ELF declares. For GQA
+    (kv_dim<emb_dim) that resolves to Q only (arg13); for MHA (kv_dim==emb_dim)
+    to Q,K,V (arg13,14,15). Hardcoding "Q only" was the GQA assumption that
+    produced zero K/V at kv_dim==emb_dim.
+    """
+    from llama32_1b.gemm_builder import gemm_registry_config
+
+    q_spec = gemm_registry_config(seq_len, emb_dim, emb_dim, "bf16", "high")
+    k_spec = gemm_registry_config(seq_len, emb_dim, kv_dim, "bf16", "high")
+    v_spec = gemm_registry_config(seq_len, emb_dim, kv_dim, "bf16", "high")
+    arrays, inter = [], set()
+    nxt = 13
+    for spec, cols in ((q_spec, emb_dim), (k_spec, kv_dim), (v_spec, kv_dim)):
+        if spec["needs_f32_scratch"]:
+            arrays.append(np.zeros((seq_len, cols), dtype=np.float32))
+            inter.add(nxt)
+            nxt += 1
+    return arrays, inter
+
+
 # Each kernel config is defined as a dict with:
 #   build_fn: callable that returns an MLIR module
 #   backend_kwargs: dict for XRTBackend constructor
@@ -275,14 +302,17 @@ def run_transformer_block(
             np.zeros((seq_len, emb_dim), dtype=bfloat16),  # q_roped_buf
             np.zeros((seq_len, kv_dim), dtype=bfloat16),  # k_roped_buf
         ]
-        # Q is fused-cast (registry best.high for the large Q shape) → needs an f32
-        # C-scratch arg (arg13); K/V are drain (no scratch).
-        _rms_args.append(np.zeros((seq_len, emb_dim), dtype=np.float32))  # arg13 q_f32
-        _arg_cache[_rms_key] = _rms_args
-    cached_args = _arg_cache[_rms_key]
+        # Registry-driven f32 C-scratch args, in builder order (Q,K,V). One per
+        # fused-cast GEMM (indices from 13). GQA (kv_dim<emb_dim) → Q only (arg13);
+        # MHA (kv_dim==emb_dim) → Q,K,V (arg13,14,15). Mirrors the builder's own
+        # gemm_registry_config so the args match the ELF's declared scratch slots.
+        _scratch_arrays, _scratch_inter = _rms_scratch_specs(seq_len, emb_dim, kv_dim)
+        _rms_args.extend(_scratch_arrays)
+        _arg_cache[_rms_key] = (_rms_args, _scratch_inter)
+    cached_args, _scratch_inter = _arg_cache[_rms_key]
     cached_args[0] = np.asarray(x_bf16, dtype=bfloat16).reshape(seq_len, emb_dim)
 
-    _rms_inter = {2, 4, 6, 8, 11, 12, 13}
+    _rms_inter = {2, 4, 6, 8, 11, 12} | _scratch_inter
     results = cache.load_and_run(
         "rms_gemms_rope",
         _rms_gemms_rope_run_backend(),
@@ -453,10 +483,14 @@ def preload_prefill_weights(weights, config, cache, seq_len, rope_lut_bf16):
             np.zeros((seq_len, emb_dim), dtype=bfloat16),  # arg11
             np.zeros((seq_len, kv_dim), dtype=bfloat16),  # arg12
         ]
-        # Q is fused-cast → arg13 q_f32 scratch.
-        rms_args.append(np.zeros((seq_len, emb_dim), dtype=np.float32))  # arg13 q_f32
-        _arg_cache[f"rms_gemms_rope_L{layer_idx}"] = rms_args
-        _rms_inter_warm = {2, 4, 6, 8, 11, 12, 13}
+        # Registry-driven f32 C-scratch (Q,K,V order). Allocates the per-layer BO
+        # set with the SAME arg count the block uses at inference time — must match
+        # run_transformer_block above or the bo_key-reused BO set would be missing
+        # MHA's K/V scratch (the original GQA-hardcoded bug). GQA → 1; MHA → 3.
+        scratch_arrays, scratch_inter = _rms_scratch_specs(seq_len, emb_dim, kv_dim)
+        rms_args.extend(scratch_arrays)
+        _arg_cache[f"rms_gemms_rope_L{layer_idx}"] = (rms_args, scratch_inter)
+        _rms_inter_warm = {2, 4, 6, 8, 11, 12} | scratch_inter
         cache.load_and_run(
             "rms_gemms_rope",
             _rms_gemms_rope_run_backend(),
